refactor: Refactor oauth
